Support DataURI files as encoded by symfony normalizer. Additional support for url encoded text files
