+++ features/assets/files/test_file.txt
@@ -1,0 +1,1 @@
+test text file
